Refactor resume module
