+++ src/menu/resume/__init__.py
@@ -1,0 +1,10 @@
+"""
+src/menu/resume/__init__.py
+
+Resume menu package: exports the menu entry point and flow helpers.
+"""
+
+from .menu import view_resume_items
+from .flow import _handle_create_resume, _handle_view_existing_resume
+
+__all__ = ["view_resume_items", "_handle_create_resume", "_handle_view_existing_resume"]
