+++ tests/test_resume_snapshot.py
@@ -5,8 +5,8 @@ from datetime import datetime, UTC
 import pytest
 
 from src.db import init_schema, save_project_summary, list_resumes, get_resume_snapshot, insert_resume_snapshot
-from src.menu import resume_menu as resume_mod
-from src.menu import resume_flow
+from src.menu.resume import menu as resume_menu
+from src.menu.resume import flow as resume_flow
 
 
 def _make_summary(project_name: str, project_type: str = "code", project_mode: str = "individual") -> str:
@@ -50,7 +50,7 @@ def test_create_resume_uses_top_five(monkeypatch):
     monkeypatch.setattr("builtins.input", lambda _: "")
     monkeypatch.setattr(resume_flow, "collect_project_data", fake_collect_project_data)
 
-    resume_mod._handle_create_resume(conn, user_id, "TestUser")
+    resume_flow._handle_create_resume(conn, user_id, "TestUser")
 
     snapshots = list_resumes(conn, user_id)
     assert len(snapshots) == 1
@@ -74,7 +74,7 @@ def test_create_resume_no_summaries(monkeypatch):
     # No summaries seeded
     monkeypatch.setattr("builtins.input", lambda _: "")
 
-    resume_mod._handle_create_resume(conn, user_id, "TestUser")
+    resume_flow._handle_create_resume(conn, user_id, "TestUser")
 
     snapshots = list_resumes(conn, user_id)
     assert snapshots == []
@@ -94,7 +94,7 @@ def test_view_existing_resume_lists_and_renders(monkeypatch, capsys):
     inputs = iter(["1"])
     monkeypatch.setattr("builtins.input", lambda _: next(inputs))
 
-    handled = resume_mod._handle_view_existing_resume(conn, user_id)
+    handled = resume_flow._handle_view_existing_resume(conn, user_id)
 
     captured = capsys.readouterr().out
     assert "MyResume" in captured
@@ -107,7 +107,7 @@ def test_view_existing_resume_no_resumes(capsys):
     init_schema(conn)
     user_id = 1
 
-    handled = resume_mod._handle_view_existing_resume(conn, user_id)
+    handled = resume_flow._handle_view_existing_resume(conn, user_id)
     captured = capsys.readouterr().out
     assert "No saved resumes yet" in captured
     assert handled is False
@@ -129,7 +129,7 @@ def test_create_resume_with_fewer_than_five_projects(monkeypatch):
     monkeypatch.setattr("builtins.input", lambda _: "")
     monkeypatch.setattr(resume_flow, "collect_project_data", fake_collect_project_data)
 
-    resume_mod._handle_create_resume(conn, user_id, "TestUser")
+    resume_flow._handle_create_resume(conn, user_id, "TestUser")
 
     snaps = list_resumes(conn, user_id)
     assert len(snaps) == 1
